Create Computer_Mice_Program.c
Till now we have seen only static output on the black window in C/C++ without any peripheral device interaction(like a mouse). Here static means that interaction with the output screen through the mouse to run a dynamic event or task. The goal will be to make the pointer of the mouse visible on our output screen, through which it can see any new event when the mouse is clicked on the same output window.